commit16

+++ Projet1/app/src/main/java/ca/qc/cgodin/projet1/SuccursaleDao.kt
@@ -16,7 +16,7 @@ interface SuccursaleDao {
     @Query("SELECT * from succursale_table WHERE Aut=(:Aut) ORDER BY Ville ASC")
     fun getSuccursalesFromAut(Aut: Int): LiveData<List<Succursale>>
     @Query("SELECT * FROM succursale_table WHERE Aut=(:Aut) AND Ville=(:Ville)")
-    fun getSuccursale(Aut: Int, Ville: String): LiveData<Succursale?>
+    fun getSuccursale(Aut: Long, Ville: String): LiveData<Succursale?>
     @Delete
     fun delete(succursale: Succursale)
     @Insert(onConflict = OnConflictStrategy.IGNORE)
